+++ detprocess/process/__init__.py
@@ -1,4 +1,4 @@
-from .process_features import *
-from .process_triggers import *
+from .features import *
+from .triggers import *
 from .processing_data import *
 from .randoms import *
