+++ detprocess/process/triggers.py
